requirements.txt
Django==1.8.3
django-pyodbc-azure==1.8.3.0
django-select2-forms==1.1.18
django-tastypie==0.12.2
pyodbc==3.0.10
python-dateutil==2.4.2
python-mimeparse==0.1.4
pytz==2015.4
six==1.9.0
